try to fix performance issue with render queries
Even when disabled , in KX_KetsjiEngine::BeginFrame(), this:

for (RAS_Query& query : m_renderQueries) {
			query.Begin();
					}
is called. Maybe it's the origin of performance issue for Akira_san

Else I noticed that there was no versioning.In some old files, queries
are enabled by default (I guess it's a flag conflict as the order of
debug properties changed). I didn't do versioning because there was no
property ShawRenderQueries in DNA struct but maybe it could be done